refactor: auth API 구조 변경 및 불필요한 파일 삭제, 타입 정의 통합
